+++ packages/ibm-products/.storybook/main.js
@@ -10,9 +10,9 @@ import { dirname, join, resolve } from 'path';
 import remarkGfm from 'remark-gfm';
 
 const storyGlobs = [
-  '../../ibm-products/src/**/*.stories.*',
   '../src/**/*.stories.*',
-  '../src/**/*.mdx',
+  '../../core/src/**/*.stories.*',
+  '../../core/src/**/*.mdx',
   '../../../examples/carbon-for-ibm-products/example-gallery/src/example-gallery.stories.js',
 ];