+++ packages/ibm-products/src/components/Carousel/Carousel.docs-page.js
@@ -8,7 +8,7 @@
 import React from 'react';
 import { StoryDocsPage } from '../../global/js/utils/StoryDocsPage';
 
-import * as stories from './Carousel.stories';
+import * as stories from './Carousel.internal.stories';
 
 const DocsPage = () => (
   <StoryDocsPage