+++ packages/ibm-products/src/components/FilterPanel/FilterPanel.stories.jsx
@@ -24,7 +24,7 @@ import {
 import mdx from './FilterPanel.mdx';
 
 import styles from './_storybook-styles.scss?inline';
-import { Annotation } from '../../../../core/.storybook/Annotation';
+import { Annotation } from '../../../.storybook/Annotation';
 
 const storyClass = 'filter-panel-stories';