+++ config/storybook-addon-carbon-theme/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.15.0-rc.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/storybook-addon-theme@2.11.0-rc.0...@carbon/storybook-addon-theme@2.15.0-rc.0) (2025-05-05)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+# [2.14.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/storybook-addon-theme@2.14.0-rc.0...@carbon/storybook-addon-theme@2.14.0) (2025-04-30)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
 # [2.14.0-rc.0](https://github.com/carbon-design-system/ibm-products/compare/@carbon/storybook-addon-theme@2.11.0-rc.0...@carbon/storybook-addon-theme@2.14.0-rc.0) (2025-04-21)
 
 **Note:** Version bump only for package @carbon/storybook-addon-theme